--- app/controllers/graphql_devise/concerns/set_user_by_token.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module GraphqlDevise
-  module Concerns
-    module SetUserByToken
-      extend ActiveSupport::Concern
-
-      included do
-        include DeviseTokenAuth::Concerns::SetUserByToken
-        include AdditionalControllerMethods
-      end
-    end
-  end
-end
