+++ lib/generators/graphql_devise/install_generator.rb
@@ -24,7 +24,7 @@ module GraphqlDevise
       end
       gsub_file(
         'app/controllers/application_controller.rb',
-        'GraphqlDevise::Concerns::SetUserByToken',
+        'GraphqlDevise::SetUserByToken',
         'DeviseTokenAuth::Concerns::SetUserByToken'
       )
 
@@ -62,7 +62,7 @@ module GraphqlDevise
       gsub_file(
         'app/controllers/application_controller.rb',
         /^\s+include DeviseTokenAuth::Concerns::SetUserByToken/,
-        '  include GraphqlDevise::Concerns::SetUserByToken'
+        '  include GraphqlDevise::SetUserByToken'
       )
     end
