+++ lib/graphql_devise/concerns/set_user_by_token.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module GraphqlDevise
+  module SetUserByToken
+    extend ActiveSupport::Concern
+
+    included do
+      include DeviseTokenAuth::Concerns::SetUserByToken
+      include AdditionalControllerMethods
+    end
+  end
+end
